add a polyfill for str::split_once
This is not a Good Idea, but rather an experiment to see what it's like
to shim in wrappers based on compiler features. It's actually not too
ugly, besides having to allow the unstable_name_collisions lint, which
I can only make work as an inner attribute on the whole file so it can't
be grouped with the rest of the polyfill code.

Too bad I need another dependency, but at least rustversion is
a standalone crate, despite being a proc macro.

6/10: interesting, but inelegant. Revert this in a couple month when
Rust 1.52 stable releases.